+++ client/private/js/src/modules/windowManager/window.js
@@ -1,4 +1,4 @@
-﻿/*
+/*
  *
  * Copyright (c) 2018 João Pedro Martins Neves (shivayl) - All Rights Reserved.
  *
@@ -10,8 +10,13 @@
 // TODO: Add the z-index of each each window.
 
 class Window {
-  constructor(processId, title, content) {
-
+  /**
+   * 
+   * @param { string } processId
+   * @param { string } title
+   * @param { string } content
+   */
+  constructor( processId, title, content ) {
     this.id = `${Window.idPrefix}${processId}`;
     this.title = title;
     this.content = content;
@@ -24,7 +29,7 @@ class Window {
   get element() { return document.getElementById( this.id ); }
   static get idPrefix() { return 'win-'; }
 
-  get template() {
+  get windowTemplate() {
     return `
       <article class="window-manager grid-y resizable selected-win" id="${this.id}">
         <header class="toolbar">
@@ -65,8 +70,20 @@ class Window {
       </div>`;
   }
 
+  static appStoreAppWindowTemplate( url ) {
+    return `
+      <iframe
+        title=""
+        src="${url}"
+        allowpaymentrequest="false"
+        sandbox="allow-scripts"
+      >
+      </iframe>
+    `;
+  }
+
   init() {
-    document.getElementById( 'window-manager-container' ).innerHTML += this.template;
+    document.getElementById( 'window-manager-container' ).innerHTML += this.windowTemplate;
   }
 
   kill() {