+++ client/private/js/src/systemApps/appStore/appStore.model.js
@@ -17,7 +17,7 @@ class AppStoreModel {
     try {
       /** @type { Response } */
       const res = await HttpClient.get( API_ROOT_PATH + 'app-store?lastId=' + id + '&limit=6' );
-      return !res.ok() ? Notifications.errorToast( res.josn() ) : await res.json();
+      return !res.ok() ? Notifications.errorToast( await res.json() ) : await res.json();
 
     } catch ( e ) {
       console.debug( e );