+++ server/portfolio-os/routes/apiIndex.js
@@ -30,6 +30,7 @@ router.delete( '/user/profile/skills/:skillId', usersRoute.deleteSkill );
 router.post( '/user/profile/links', usersRoute.addLink );
 router.put( '/user/profile/links/:linkId', usersRoute.updateLink );
 router.delete( '/user/profile/links/:linkId', usersRoute.deleteLink );
+router.get( '/user/installed-apps', usersRoute.getInstalledApps );
 
 // #endregion